refactor: move Sentence and Verb to dedicated subfolders for better organization
